grok5 tried to put code mods into downloadable file artifacts which would have been great for avoiding the markup rendering problem in the chat panel, except for the fact that grok5 lies about the downloadable file artifact actually existing.  This is apparently becuase the temp file system it uses gets overwritten during its thinking process.
Also, Sonnet 4 continually screws up its code mods by putting them in the code panel where it
tries to "edit" them if there are any mistakes -- sometimes in the middle of thinking about
it -- which leaves things in a mess.  Getting things put in a non-editable code section of
the chat panel can avoid the worst of this behavior.